imgsensors: mt9m114_mipi_raw: Enable NightMode
On stock ASUS kernel, this is disabled (confirmed by reverse engineering).
Intead, on stock this function, as well as AE (auto exposure), and AWB (auto white balance), and others, are handled by a custom chip, the 'iCatch SPCA7002A'

This custom kernel has no support for the iCatch SPCA7002A, the driver for MediaTek was ported by ASUS from QCOMM, hence, closed source.

So, to compensate for the lack of this chip, enable 'the default' image sensor NightMode, this makes a huge difference on low light conditions.